backend/requirements.txt
pydantic==2.12.5
fastapi==0.123.9
python-multipart==0.0.20
uvicorn==0.38.0
psycopg2-binary==2.9.11
sqlalchemy==2.0.36
sqlalchemy-cockroachdb==2.0.3
python-dotenv==1.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1